{"types":"C","runs":11}

+++ Problem_30.java
@@ -7,7 +7,7 @@ public class Problem_30 extends ConsoleProgram {
 	public void run () {
 		int n = readInt("enter the integer: ");
 		double m = 5.4;
-		int x = (double) Math.min(n, m);
+		double x = (double) Math.min(n, m);
 		println("x: " + x);
 
 	}
